add a `Manuscript.slug` field
Three db migrations are required here: the first adds the field allowing `NULL` values; the second manually fills in a value for each `Manuscript` object; and the third adds the `NOT NULL`  and `UNIQUE` constraints back.

The over-riding of `Manuscript.save` and the addition of `ManuscriptAdmin.prepopulated_fields` should ensure that all new manuscripts are covered.

The slug is added to the admin interface for easy referenceability.

IMPORTANT: these migrations mean that new databases cannot be populated from old `dumpdata` dumps (but there will be more changes that break compatibility with these dumps soon anyway, and the current dump linked in the repo `README.md` doesn't work anyway...).